% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluates.R
\name{evalPolynomial}
\alias{evalPolynomial}
\title{Evaluate a polynomial}
\usage{
evalPolynomial(poly, coefficients = NULL, powers = NULL, X)
}
\arguments{
\item{poly}{a polynomial given as a list of two elements: the powers in an 
integer matrix and the coefficients of the monomials composing the polynomial, 
like a \code{\link[spray]{spray}} object}

\item{coefficients, powers}{an alternative way to input the polynomial, to use 
with missing \code{poly}}

\item{X}{a matrix whose rows are the evaluation points}
}
\value{
A numeric vector, the values of the polynomial at the points given by 
\code{X}.
}
\description{
Evaluation of a multivariate polynomial at some values of the 
variables.
}
